// medusa/.medusa/types/query-entry-points.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: Date | string; output: Date | string; }
  JSON: { input: Record<string, unknown>; output: Record<string, unknown>; }
};

export type SalesChannel = {
  __typename?: 'SalesChannel';
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  description: Maybe<Scalars['String']['output']>;
  is_disabled: Scalars['Boolean']['output'];
  created_at: Scalars['DateTime']['output'];
  metadata: Maybe<Scalars['JSON']['output']>;
  updated_at: Scalars['DateTime']['output'];
  deleted_at: Maybe<Scalars['DateTime']['output']>;
  carts: Maybe<Array<Maybe<Cart>>>;
  orders: Maybe<Array<Maybe<Order>>>;
  products_link: Maybe<Array<Maybe<LinkProductSalesChannel>>>;
  api_keys_link: Maybe<Array<Maybe<LinkPublishableApiKeySalesChannel>>>;
  publishable_api_keys: Maybe<Array<Maybe<ApiKey>>>;
  locations_link: Maybe<Array<Maybe<LinkSalesChannelStockLocation>>>;
  stock_locations: Maybe<Array<Maybe<StockLocation>>>;
};

export type ApiKeyTypeEnum =
  | 'publishable'
  | 'secret';

export type ApiKey = {
  __typename?: 'ApiKey';
  id: Scalars['ID']['output'];
  token: Scalars['String']['output'];
  salt: Scalars['String']['output'];
  redacted: Scalars['String']['output'];
  title: Scalars['String']['output'];
  type: ApiKeyTypeEnum;
  last_used_at: Maybe<Scalars['DateTime']['output']>;
  created_by: Scalars['String']['output'];
  revoked_by: Maybe<Scalars['String']['output']>;
  revoked_at: Maybe<Scalars['DateTime']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at: Maybe<Scalars['DateTime']['output']>;
  sales_channels_link: Maybe<Array<Maybe<LinkPublishableApiKeySalesChannel>>>;
  sales_channels: Maybe<Array<Maybe<SalesChannel>>>;
};
// ...
export type StockLocation = {
  __typename?: 'StockLocation';
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  metadata: Maybe<Scalars['JSON']['output']>;
  address_id: Scalars['ID']['output'];
  address: Maybe<StockLocationAddress>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at: Maybe<Scalars['DateTime']['output']>;
  fulfillment_provider_link: Maybe<Array<Maybe<LinkLocationFulfillmentProvider>>>;
  fulfillment_providers: Maybe<Array<Maybe<FulfillmentProvider>>>;
  fulfillment_set_link: Maybe<Array<Maybe<LinkLocationFulfillmentSet>>>;
  fulfillment_sets: Maybe<Array<Maybe<FulfillmentSet>>>;
  sales_channels_link: Maybe<Array<Maybe<LinkSalesChannelStockLocation>>>;
  sales_channels: Maybe<Array<Maybe<SalesChannel>>>;
};
// ...
export type Cart = {
  __typename?: 'Cart';
  id: Scalars['ID']['output'];
  region_id: Maybe<Scalars['String']['output']>;
  customer_id: Maybe<Scalars['String']['output']>;
  sales_channel_id: Maybe<Scalars['String']['output']>;
  email: Maybe<Scalars['String']['output']>;
  currency_code: Scalars['String']['output'];
  locale: Maybe<Scalars['String']['output']>;
  metadata: Maybe<Scalars['JSON']['output']>;
  completed_at: Maybe<Scalars['DateTime']['output']>;
  shipping_address_id: Maybe<Scalars['String']['output']>;
  shipping_address: Maybe<Address>;
  billing_address_id: Maybe<Scalars['String']['output']>;
  billing_address: Maybe<Address>;
  items: Array<Maybe<LineItem>>;
  credit_lines: Array<Maybe<CreditLine>>;
  shipping_methods: Array<Maybe<ShippingMethod>>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at: Maybe<Scalars['DateTime']['output']>;
  customer: Maybe<Customer>;
  region: Maybe<Region>;
  sales_channel: Maybe<SalesChannel>;
  payment_collection_link: Maybe<LinkCartPaymentCollection>;
  payment_collection: Maybe<PaymentCollection>;
  cart_link: Maybe<Array<Maybe<LinkCartPromotion>>>;
  promotions: Maybe<Array<Maybe<Promotion>>>;
  order_link: Maybe<LinkOrderCart>;
  order: Maybe<Order>;
};
// ...
export type Order = {
  __typename?: 'Order';
  id: Scalars['ID']['output'];
  version: Scalars['Int']['output'];
  order_change: Maybe<OrderChange>;
  status: OrderStatus;
  region_id: Maybe<Scalars['String']['output']>;
  customer_id: Maybe<Scalars['String']['output']>;
  display_id: Maybe<Scalars['String']['output']>;
  custom_display_id: Maybe<Scalars['String']['output']>;
  sales_channel_id: Maybe<Scalars['String']['output']>;
  email: Maybe<Scalars['String']['output']>;
  currency_code: Scalars['String']['output'];
  shipping_address: Maybe<OrderAddress>;
  billing_address: Maybe<OrderAddress>;
  items: Maybe<Array<Maybe<OrderLineItem>>>;
  shipping_methods: Maybe<Array<Maybe<OrderShippingMethod>>>;
  transactions: Maybe<Array<Maybe<OrderTransaction>>>;
  summary: Maybe<OrderSummary>;
  metadata: Maybe<Scalars['JSON']['output']>;
  canceled_at: Maybe<Scalars['DateTime']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  original_item_total: Scalars['Float']['output'];
  original_item_subtotal: Scalars['Float']['output'];
  original_item_tax_total: Scalars['Float']['output'];
  item_total: Scalars['Float']['output'];
  item_subtotal: Scalars['Float']['output'];
  item_tax_total: Scalars['Float']['output'];
  original_total: Scalars['Float']['output'];
  original_subtotal: Scalars['Float']['output'];
  original_tax_total: Scalars['Float']['output'];
  total: Scalars['Float']['output'];
  subtotal: Scalars['Float']['output'];
  tax_total: Scalars['Float']['output'];
  discount_total: Scalars['Float']['output'];
  discount_tax_total: Scalars['Float']['output'];
  gift_card_total: Scalars['Float']['output'];
  gift_card_tax_total: Scalars['Float']['output'];
  shipping_total: Scalars['Float']['output'];
  shipping_subtotal: Scalars['Float']['output'];
  shipping_tax_total: Scalars['Float']['output'];
  original_shipping_total: Scalars['Float']['output'];
  original_shipping_subtotal: Scalars['Float']['output'];
  original_shipping_tax_total: Scalars['Float']['output'];
  raw_original_item_total: Maybe<Scalars['JSON']['output']>;
  raw_original_item_subtotal: Maybe<Scalars['JSON']['output']>;
  raw_original_item_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_item_total: Maybe<Scalars['JSON']['output']>;
  raw_item_subtotal: Maybe<Scalars['JSON']['output']>;
  raw_item_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_original_total: Maybe<Scalars['JSON']['output']>;
  raw_original_subtotal: Maybe<Scalars['JSON']['output']>;
  raw_original_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_total: Maybe<Scalars['JSON']['output']>;
  raw_subtotal: Maybe<Scalars['JSON']['output']>;
  raw_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_discount_total: Maybe<Scalars['JSON']['output']>;
  raw_discount_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_gift_card_total: Maybe<Scalars['JSON']['output']>;
  raw_gift_card_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_shipping_total: Maybe<Scalars['JSON']['output']>;
  raw_shipping_subtotal: Maybe<Scalars['JSON']['output']>;
  raw_shipping_tax_total: Maybe<Scalars['JSON']['output']>;
  raw_original_shipping_total: Maybe<Scalars['JSON']['output']>;
  raw_original_shipping_subtotal: Maybe<Scalars['JSON']['output']>;
  raw_original_shipping_tax_total: Maybe<Scalars['JSON']['output']>;
  customer: Maybe<Customer>;
  region: Maybe<Region>;
  sales_channel: Maybe<SalesChannel>;
  cart_link: Maybe<LinkOrderCart>;
  cart: Maybe<Cart>;
  fulfillment_link: Maybe<Array<Maybe<LinkOrderFulfillment>>>;
  fulfillments: Maybe<Array<Maybe<Fulfillment>>>;
  payment_collections_link: Maybe<LinkOrderPaymentCollection>;
  payment_collections: Maybe<Array<Maybe<PaymentCollection>>>;
  promotion_link: Maybe<Array<Maybe<LinkOrderPromotion>>>;
  promotions: Maybe<Array<Maybe<Promotion>>>;
  promotion: Maybe<Array<Maybe<Promotion>>>;
};
// ...
export type LinkProductSalesChannel = {
  __typename?: 'LinkProductSalesChannel';
  product_id: Scalars['String']['output'];
  sales_channel_id: Scalars['String']['output'];
  product: Maybe<Product>;
  sales_channel: Maybe<SalesChannel>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt: Maybe<Scalars['String']['output']>;
};
// ...
export type LinkPublishableApiKeySalesChannel = {
  __typename?: 'LinkPublishableApiKeySalesChannel';
  publishable_key_id: Scalars['String']['output'];
  sales_channel_id: Scalars['String']['output'];
  api_key: Maybe<ApiKey>;
  sales_channel: Maybe<SalesChannel>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt: Maybe<Scalars['String']['output']>;
};

export type LinkRegionPaymentProvider = {
  __typename?: 'LinkRegionPaymentProvider';
  region_id: Scalars['String']['output'];
  payment_provider_id: Scalars['String']['output'];
  region: Maybe<Region>;
  payment_provider: Maybe<PaymentProvider>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt: Maybe<Scalars['String']['output']>;
};

export type LinkSalesChannelStockLocation = {
  __typename?: 'LinkSalesChannelStockLocation';
  sales_channel_id: Scalars['String']['output'];
  stock_location_id: Scalars['String']['output'];
  sales_channel: Maybe<SalesChannel>;
  location: Maybe<StockLocation>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt: Maybe<Scalars['String']['output']>;
};